int にキャストしないでそのまま出力するようにした。

+++ core/common/servent.cpp
@@ -2502,7 +2502,7 @@ int Servent::serverProc(ThreadInfo *thread)
         if (!sv->sock)
             throw StreamException("Server has no socket");
 
-        sys->setThreadName(thread, String::format("LISTEN %d", (int) sv->sock->host.port));
+        sys->setThreadName(thread, String::format("LISTEN %hu", sv->sock->host.port));
 
         sv->setStatus(S_LISTENING);
